refactor(app): move input logic from main to InputView Object
